Return tracking number to the client
This originally started with a tracking number sent by a client (middle tier) but, a tracking number can be created by the endpoint. The user server is set to use the tracking number, has to see it again, when doing ops. So, the client has to learn of it, save it and use it.